Fix CI (#6692)
CI on `main` is failing on publishing snapshot, because javadoc is
emitting a warning and we fail on warnings. See code comment for more
explanation.

I tried pointing to
`https://docs.oracle.com/en/java/javase/11/docs/api/`, but that gives a
different error:

> warning: The code being documented uses packages in the unnamed
module, but the packages defined in
https://docs.oracle.com/en/java/javase/11/docs/api/ are in named
modules.